Add IDs to type handled by pool generator
To use `Delete` function in the Persistance, we need to know what should
be deleted. As unique identifier for files we will use a filename stored
on disk. We define a `Persisted[T]` wrapper for `T` that will add an ID.
The ID will be generated in `Save` function and added to the wrapper
type and used in `Delete` function to remove the property.