Less frequent signing requests
A signing request every 100 block is quite often giving the fact that
the signing protocol takes 73 blocks and can be retried multiple times.
Multiple simultaneous runs make debugging harder and consume a lot
of resources which is a pain on local testing environments that often
runs several clients on one machine. To facilitate testing, we make
the signing requests less frequent and trigger them every 200 block.