Fetch TSS pre-params after symmetric keys are generated
Here we delay the moment when TSS pre-params are fetched from the pool.
So far, we fetched them at the very beginning of tECDSA DKG which could
drain the pool too early when, for example, some members are inactive
from the start. This change moves the moment when pre-params are fetched
to occur just after symmetric keys are generated and actual TSS is about
to be started. Such a logic should preserve the pre-params up to the moment
when two first phases (ephemeral key exchange and symmetric key generation)
are done successfully so the chances that the group is healthy are higher.
That should result in wiser pre-params management.