Do not start tECDSA DKG if TSS pre-params pool is too small
Here we add a check of the TSS pre-params pool size upon tECDSA DKG start. So
far, the DKG started regardless of the TSS pre-params pool size which resulted
in pre-params waste and DKG execution error in case the pool was too small.
This change allows to graceful abort the DKG execution and preserve existing
pre-params by the given node if their pool turns out to be too small to perform
a successful DKG run.