Do not join the sortition pool before enough TSS pre-parameters are generated
Here we leverage the `sortition.JoinPolicy` introduced recently to enforce
the sufficient size of the DKG pre-params pool before joining the sortition
pool. Following this courtesy rule will make the network healthier as we
increase the chance that only operable nodes join the sortition pool and
are selected for protocols executions.