Use `abi.encode` instead of `abi.encodePacked`
Solidity documentation says:

If you use keccak256(abi.encodePacked(a, b)) and both a and b are dynamic types,
it is easy to craft collisions in the hash value by moving parts of a into b and
vice-versa. More specifically, abi.encodePacked("a", "bc") ==
abi.encodePacked("ab", "c"). If you use abi.encodePacked for signatures,
authentication or data integrity, make sure to always use the same types and
check that at most one of them is dynamic. Unless there is a compelling reason,
abi.encode should be preferred.

During the course of the security review, for `EcdsaDkgValidator`, we stated
that:

Solidity does not provide bytes64 type so the group public key is passed as
bytes with the length enforced to 64 by the validation code. This way, both
result.groupPublicKey and startBlock have a fixed length and only
result.misbehavedMembersIndices is of variable-length. Thus, there is no risk of
collision by using dynamic types.

For `EcdsaInactivity` this is not a problem because every inactivity claim
has a unique nonce.

Auditors stated that the length-checks are not enforced in `validateSignatures`
(which is a public function), and thereby the issue depends on the
implementation of contracts interacting with `EcdsaDkgValidator`.

The current implementation of `WalletRegistry` does not use `validateSignatures`
directly but we can replace `abi.encodePacked` with `abi.encode` for potentially
safer code long-term given the gas difference is small, about 6k of gas units
increase for `notifyOperatorInactivity` and 2k of gas units decrease for
`challengeDkgResult`.

The same change is applied to the random beacon.