Prepare Dockerfile to build binaries
We want to release the keep-client as docker image and binaries.

To build the binaries we need bunch of tools like protobuf, code
generators, etc. It's hard to setup the build environment in CI, so we
should use what we already have in Docker.

We reuse the generated sources that are used for building the docker
image (`build-sources`) and build binaries with them. In `build-bins`
target. To get the binaries out of the image we define a `output-bins`
target that will be used to output the files included in the image to a
directory with `docker buildx build` command.