Document deployments and contracts addresses
We need one source of truth when it comes to listing deployment info and
addresses of the contracts on different environments. We create a separate doc
that would hold that information and would be linked to from other documents
referencing its content. Thanks to this, after new deployment we'll need to
update the deployment info only in one place.