Append chain ID to signature schemes
During the course of the security review, auditors noticed that we do not append
chain ID to the signature hash, and as a result, if there is a fork of Ethereum’s
mainnet after the EcdsaInactivity contract’s creation, or if we upgrade the
contract to a new version, the DKG signatures will be valid on both forks.

This partially true because the same wallet must not be reused between
independent state forks because doing so would allow attackers to try to replay
wallet Bitcoin signatures. The same Bitcoin can not be spent on two state forks
so we always require Bridge and WalletRegistry contracts on any other chain to
be different.

Even if a state fork occurred, the value of the token on the state fork would
require the network and DAO recognizing it as a valid state fork. If the network
generates a valid accusation against state fork A, it is only valid against
contracts in fork B if those contracts are considered canonical. However, we
require that contracts on any other chains be different, and the network will
not recognize the same contracts on other chains.

`walletPubKey` which is a part of the input used to generate the hash for
signing is generated using a threshold ECDSA off-chain protocol and the
probability of generating two threshold ECDSA wallets with the same public key
is negligible.

To address the problem of temporary soft-forks of Ethereum mainnet, the
off-chain client needs to wait for enough confirmations before executing any
action based on the Ethereum state update received.

The rule of not reusing the wallet between state forks must not be violated.
If this rule is not violated, the signatures may not be replayed between forks.
If this rule is violated, worse things can happen than replying inactivity claim
signature.

However, given that adding the chain ID to the signature hash is pretty much
transparent in gas, we added it.

The same change has been applied to the random beacon.

I have been also considered adjusting the signature scheme to follow EIP-712.

From EIP-712:

The domain separator prevents collision of otherwise identical structures. It is
possible that two DApps come up with an identical structure like
`Transfer(address from,address to,uint256 amount)` that should not be compatible.
By introducing a domain separator the DApp developers are guaranteed that there
can be no signature collision.

In our case, the wallet must not sign anything else than Bitcoin transactions or
inactivity claims. Otherwise, the wallet may be accused for a fraud. Hence,
participating of the wallet in dApps is not possible and the domain separator is
not needed.