Configure default environment only if value not defined
Previously when a make command was executed with empty environment
`ENVIRONMENT=` the value wasn't being set to the default but left empty.
This was problematic. With `ifndef` we will set the default also when a
value is left empty.

It simplifies the CI configuration, as if we build with
`ENVIRONMENT=${{ github.event.inputs.environment }}` and the value is
not available it will be set to default by the Makefile.