Add `relese` target to `all`
The `release` target requires go code generation to happen, add it to
the `all` target to have the complete build process under `all`.

But we don't add the prerequisited to `release` target as it's used in
docker image when layers at the beginign run `make generate` and then
it's used when running `make release`. We don't want to duplicate the
work.