Mock WalletRegistry with RandomBeacon
THIS IS A WORKAROUND ONLY FOR THE EARLY STAGE OF THE CHAOSNET

Use `WalletRegistryContractName` instead of `RandomBeaconContractName`.
DKG for the WalletRegistry depends on the RandomBeacon group creation.
Currently the client doesn't publish a generated group to the chain
as it works against a mocked chain implementation. Without a Beacon group
published to the chain, the WalletRegistry's DKG cannot start. As a workaround
for the first stage of the Chaosnet we use the RandomBeacon's address,
as the client only wants to get to the sortition pool to select a group.