Stop pill as a workaround for no consensus integration
`StopPill` is a temporary workaround for a missing chain integration. When a
group member is not selected for the current attempt of key generation or
signing and there is no other member from the same client selected for the
protocol execution, the member does not know what was the result of the protocol
execution and if it completed or not. In other words, `defer cancelLoopCtx()`
does not get executed and the member will stay hung on the block waiter, waiting
for their turn. The `StopPill` is sent via broadcast channel on a successful
protocol execution and tells all members waiting for their turn in the retry
loop to stop because the result was produced.