Refresh go.mod
The changes in go.mod are not a result of the work on stop pill or retry
loop. These changes should be made earlier, when working on the context
logging in https://github.com/keep-network/keep-core/pull/3267 and
having our Go files import ipfs/go-log/v2 and zap.

The go vet is complaining so I am updating go.mod now.