Call metric Observe functions on a pointer
This is a preparation for exposing metrics registry to applications.
Calling functions on a pointer feels better anyway than passing the pointer
as a parameter.

I have also fixed a but in start.go: application diagnostics are not registered
if diagnostics are not configured. If diagnostics are not configured,
diagnostics registry pointer is nil.