Beta operator sortition pool policy for tbtc and random beacon
BetaOperatorPolicy is a JoinPolicy implementation checking chaosnet and beta
operator status. If chaosnet has been deactivated, the operator is allowed to
join the pool. If chaosnet is active and the operator is beta operator, the
operator is allowed to join the pool. If chaosnet is active and the operator is
not beta operator, the operator is not allowed to join the pool.

The BetaOperatorPolicy is applied both to the random beacon and tbtc.