Configure only public address in announced addresses
The nodes shouldn't configure an internal address for announced
addresses as the address is not rechable and is considered incorrect
configuration for the network. We configure the announced addresses
only if we want to expose a DNS address for a bootstrap nodes. The rest
of nodes is resolving the addresses with libp2p to the public addresses
correctly by default.