Merge diagnostics and metrics into a single clientinfo registry
Previously, we were exposing diagnostics and metrics on separate ports but the
reality was that we had two HTTP servers exposing exactly the same content: the
diagnostics port could be queried for both /diagnostics and /metrics and the
metrics port could be queried for both /metrics and /diagnostics. This change
simplifies the setup into a single clientinfo package and registry exposing
metrics and diagnostics.

From the operator's perspective the only change is the config file and the fact
one less port is used by the client.

This commit is a draft, unfinished work. Unit tests are not passing and the
functionality was not tested.