Add session ID to the beacon relay entry messages
So far, the messages carrying signature shares specific to a relay entry
request were identified only by the sender ID. That turned to be problematic
if multiple relay entry requests target the same group in a short period of
time. Due to network retransmissions, old shares messages referring to a
previous entry request may be handled by members listening for share
messages specific for the currently processed entry request. If that scenario
happens, group members reject the wrong shares related to previous entry and
produce log warnings. Although that case does not harm the protocol execution
at the moment, it is generally error-prone and may lead to unpredictable
consequences in the future. As remediation, we are extending the relay entry
share message definition by adding a session ID field that allows attributing
the message to a specific relay entry request. Then, the session ID can be
used by group members to filter out all messages that does not belong to the
currently processed relay entry request.