Add session ID to the beacon DKG messages
So far, the messages carrying information specific to a DKG instance were
identified only by the sender ID. That can be problematic if multiple DKG
instances occur in a short period of time. Due to network retransmissions,
old DKG messages referring to a previous DKG instance may be handled by members
listening for DKG messages specific for the currently executed DKG instance. If
that scenario happens, group members can turn into an inconsistent state and
break the protocol execution. As remediation, we are extending the DKG messages
definitions by adding a session ID field that allows attributing the messages to
a specific DKG instance. Then, the session ID can be used by group members to
filter out all messages that does not belong to the currently processed DKG
instance.