New Merkle Distribution for Sep 30th
This is the rewards distribution held on Sep 1st 2022.

As this is a cumulative Merkle distribution, the rewards here collected
are the sum of the rewards distributed in previous Merkle distributions
plus the rewards earned in the period between the last and the current
distributions.

This includes:

- Bonus rewards for whose who were staking and had an PRE operator
  confirmed before June 1st 2022 and didn't unstaked any amount of
  tokens until July 15th 2022.
- Ongoing rewards for those who were staking and had an PRE operator
  confirmed between June 1st and Sep 30th 2022.

Note: this distribution has been generated after subgraph fix
https://github.com/threshold-network/subgraph/pull/30
so the bonus rewards has been slightly increased for one staker since it
was not being taken into account that a PRE operator was confirmed
before the start of bonus rewards earning period.