Ensure the ECDSA DKG result is signed by the quorum of group members
Even if that requirement is not forced by the on-chain contract, every
ECDSA DKG participant must ensure it submits a DKG result that was signed
by the quorum of the group. This way we are making sure the protocol was
completed by a quorum of members that are ready to handle signature requests.

This change should soften the case when members fail at the last phase
of DKG (finalization round) that is not followed by any communication round
allowing for detection of such a failures. In that scenario, members who
failed the last round do not produce the key material but, in the same time,
are not excluded by other participants as misbehaved, so they can be selected
for signing normally. As result, signature attempts including those members
fail and the wallet needs more attempts in total to exclude those members
using the random algorithm and come with a successful signature. However,
thanks to this change, we now guarantee that the total number of signing-ready
members is at least equal to the `GroupQuorum` which should strongly increase
the success chance of particular signing attempts and result with a lower number
of total trials (comparing to at least `HonestThreshold` signing-ready members).

Note that the problem may still occur if the number of successful participants
is greater than the `GroupQuorum` yet less than the total number of
members participating in the last DKG attempt. The ultimate solution should
probably involve an additional communication round after the ECDSA DKG
finalization round.