WORKAROUND: Remove tbtc code generation
`tbtc` package is not yet supported in the client. The contracts were
not yet migrated on mainnet so the build process for mainnet cannot
complete. We remove the tbtc support in code generator to let the client
build for mainnet. This whole commit should be reverted when adding
integration with tbtc contracts.