Log problems with selecting group on warn level
We should consider switching this log to Errorf when the Chaosnet 0 phase is
completed and results are submitted to the chain. To let the operators join the
pool, the dev team will keep unlocking it via DKG timeout from time to time.
During the period pool is unlocked, selecting the group will not work.