Drafting rewards requirements
There are 5 requirements for a peer to receive rewards:
- Random Beacon authorized
- tBTC authorized
- Node up time 96% after a peer joins network
- Pre generated params (avg daily of 500 for now)
- Build version (yet to be implemented)

rewards-requirements.ts:
- call beacon and wallet registry contracts for authorization info
- queries Prometheus service for bootstrap peers info
- aggregate and fetches rewards requirements for each peer

Each requirement has a binary factor: 0 or 1 meaning if a given
requirement is satisfied. If any of the requirement is 0 it means no
rewards for a peer in a given interval. Up time requirement has an
additional coefficient which denotes a portion of rewards in a given
rewards interval. For example if a peer joins mid month, then only 50%
of rewards will be distributed to this peer.