Add confirmation state at the end of ECDSA DKG
So far, members' failures in the last state (finalization) of the ECDSA DKG
could remain undetected by other participants as no explicit confirmation step
was done after the result was produced. The result signature process that
occurs after the DKG finalization is able to detect such a failures to some
extent, only when the signatures count is lower than the group quorum parameter.
Undetected DKG finalization failures lead to a spurious view regarding the
wallet's state because some members considered as properly operating are
actually inactive. Such inactive members are normally selected for the signing
work despite being unable to participate in the protocol. That leads to multiple
signing failures that must be compensated by the random retry algorithm. This
extends the signature time abnormally.

To mitigate that problem, we are extending the DKG finalization state to
produce and exchange messages informing about successful finalization. We are
also introducing a new confirmation state that occurs after the finalization
and makes sure that all DKG participants selected for current DKG attempt
sent a finalization message thus were able to produce the result.