Introduce readiness announcement before a signing attempt
Here we leverage the previously introduced `tbtc.signingAnnouncementMessage`
by implementing a readiness announcement phase before each signing attempt
of the given message. This phase is meant to determine a set of active
members from which the random algorithm can safely select a 51-members subset
supposed to perform the signing attempt with a high success chance.

This change should improve the overall signing success rate by alleviating
the negative impact of inactivites. Moreover, the mechanism presented here
plays nicely with the design described in the RFC-7 and can be used to
synchronize members that must sign the right thing proposed by the coordinator.
Last but not least, thanks to announcements, operators whose clients were
restarted will have a chance to catch up the ongoing signing processes and
increase their success chance.