Fast state machine
Package faststate contains a generic state machine implementation that is
meant to be used with interactive protocols which do not require a strict
synchronization mechanism between protocol members. The synchronization is
based on an signal from each participant that they are ready to proceed to
the next step when, for example, they received all the necessary information
from all the other participants.

This approach allows for faster execution of protocols but has strict
requirements regarding the implementation of states.

Requirement 1: Context lifetime and retransmissions

The context passed to `faststate.NewMachine` must be active as long as the
result is not published to the chain or until a fixed time for the protocol
execution has not passed.

The context is used for the retransmission of messages and all protocol
participants must have a chance to receive messages from other participants.
Consider the following example: there are two participants of the protocol:
A and B, and they are executing the final state of the protocol. If the
context was canceled right after the completion of work by the participant,
without a confirmation that the result was published on-chain, we could run
into a situation when A received a message from B and exited immediately,
without giving B a chance to receive a message:

A: |-S------R-|
B:        |-S----------(...)

|- denotes when the last state starts and -| denotes when it ends. S denotes
when the given participant sent its message and R denotes when the given
participant received the message from the other. Since B initiated the last
state later than A (it could execute some time-consuming computations in the
last state), A already sent its message and now B needs to wait for the
retransmissions. B sends its message immediately upon the initiation and this
message is received by A. Since A exits and cancels the context immediately
after receiving a message from B, it is no longer retransmitting its message
and B hangs forever.

There are two solutions. One is that both A and B observe the chain and they
keep the context active until the result is published. Another is that each
member waits for some time before completing the protocol and canceling the
context for retransmissions or that the entire protocol has a fixed maximum
execution time until the timeout.

Requirement 2: Store all received messages

Since the state machine does not require strict synchronization between
participants, it is not guaranteed at which moment of the execution the rest
of the group is. If the current member is at the first state of the
execution, and all other members advanced to further states, the current
member should accept and store messages from further states "for the future"
instead of rejecting them. This can be achieved using `faststate.BaseState`
structure.