Install Go outside of `staticcheck-action`
Previously we were using `dominikh/staticcheck-action@v1.2.0` action to install
Go. Theoretically, the action should use the version of go specified in `go.mod`
file (`1.18`), but we observed that older version (`1.17`) is installed instead.
This haven't changed even when we explicitly configured the action with
`min-go-version: "1.18"` parameter. This mismatch of versions was resulting in
the tool reporting many false problems. To fix that, we decided to install Go
via dedicated `actions/setup-go@v3` action and configure
`dominikh/staticcheck-action@v1.2.0` to not run the Go installation.