Move the documentation of contracts addresses to Threshold docs
When possible, we want to reduce the number of docs storing the same
information, ideally to have a single source of truth. As the contracts
addresses got documented in the Threshold docs
(https://docs.threshold.network/extras/contract-addresses), we can reference
that documentation in Keep docs insteade of having a separate doc on Keep side.