Adjusted state package Go documentation
The goal of the changes is to merge together `faststate` and `state` packages.
Instead of two packages, we are going to have one `state` package with two
implementations of the state machine: asynchronous and synchronous.

state.SyncMachine is meant to be used with interactive protocols when
participants are expected to synchronize based on the number of blocks being
mined at the time the protocol is executing. Even if the given participant
received all the necessary information to continue the protocol, the state
machine waits with proceeding to the next step for the fixed duration of
blocks. This approach is the most optimal when the protocol may finish
successfully even if some members expected to participate in the execution
are inactive.

state.AsyncMachine is meant to be used with interactive protocols when
participants are expected to synchronize based on the messages being sent and
some participants may be slower than others. Each protocol participant, to
finish the execution, must wait for the slowest participant. This approach is
the most optimal when the protocol may finish successfully only if all
members expected to participate in the execution are actively participating.