Use recreate strategy for prometheus
Prometheus uses a storage where it adds a lockfile. It causes problems
with rolling updates as the lockfile has to be released before running a
new instance.

We have two solutions:
- stop using rolling updated
- add `--storage.tsdb.no-lockfile` argument to prometheus

For now we decided to use the simplier solution so we don't cause any
data corruption because the lockfile ignore.