Use child contexts for message sending to avoid overwriting retransmissions
We now use a global context in the state machine which is active until the
timeout block. This context is used in all `channel.Send` calls that manage
retransmissions under the hood. The context is used as key in the map that holds
the functions responsible for specific retransmissions. If we call `channel
.Send` with the same context for different messages, the latest call will
overwrite previous retransmissions with the current one. To avoid that,
we use a separate child context for each message.