Fetching authorizations for a given rewards interval
Fetched and calculated authorization for a given rewards interval. The
authorization is calculated based on the events that occured during that
interval. Ex.
- Inteval Sep1-Sep30
- Increase event Sep10 from 100k to 110k
authorization=10/30*100k+20/30*110k
If there were no events emitted, then the next event after the interval
is used to fetch the authorization that was during the
interval. If no events were emitted, it means that we can get the
current authorization which lasted for the whole rewards interval time.