Assign each retransmission request a unique id
Previously we were identifying retransmission handlers by context. This did work
well when each state for each member had a separate context in the synchronous
state machine. It did not work well though when the same context was shared for
all states in the async state machine. Given that the retransmission ticker
identified retransmission functions by context, new retransmission requests were
overwriting previous ones using the same context. We fixed it temporarily by
creating new instances of child contexts but that was not a final solution we
wanted to go with. This commit fixes the problem in a proper way (TM). Each
retransmission has a unique ID and no matter if the same or separate contexts
are used, registering new retransmission requests does not overwrite existing
ones.