Asynchronous state machine: do not block on state initiation
In the async state machine we gather all messages received, also those from the
"future" states. Blocking `Initiate` makes sense for the sync state machine,
where we may want to initiate a state before we are ready to accept messages.
For the async state machine, we want to accept messages all the time and
blocking `Initiate` is not needed. Having `Initiate` running asynchronously will
allow us to `Recv` messages to states and do not overlflow the buffer in the
state machine in case the computations take longer than expected. It should also
allow for a faster execution of states because the messages are received also at
the time of Initiating the state so we can potentially transition right after
the computations are done.

The state machine should not be asking the state if it `CanTransition` if
`Initiate` of the state has not been completed yet. This was guaranteed with a
blocking `Initiate` and has been maintained for the non-blocking `Initiate`.
I've made it clear in the documentation of the AsyncState.

Please see the next commit as a proof how this can simplify the code.