Add pre-params requirements
Pre params are calculated as an average value across all the instances for a
given operator in a given rewards interval. The avg of a total pre
params number across the instances is taken to validate the requirement,
ie. a single instance is not validated, but all of them for the same
operator.