Abstract `net.Message` payload extraction and deduplication
A common scenario occurring during the work with `net.Message` and `state
.BaseAsyncState` is extracting message payloads of given type from the
history, and ensuring their uniqueness using the sender as key. So far, this
logic was defined in the tECDSA signing protocol code but, we noticed it should
live on a higher abstraction level to avoid duplicating it across different
protocols. That said, we decided to make it generic and move to the
`pkg/state`. This way, specific protocols only need to assemble the right call
without defining the logic itself.