Identify stop pill by DKG seed or signed message
The stop pill was sent in a broadcast channel and stopped all signing and key
generation using that broadcast channel. This worked fine for key generation,
because broadcast channel name for key generation is computed from the last
block's hash. This did not work well for signing sometimes because there is one
broadcast channel used for all attempts of signing. If there were two attempts
running at the same time (e.g. multiple attempts of the previous signing and
a new signing starting), the stop pill was stopping the slower signing protocol.

With this change, each signing and DKG is uniquely identified. It was not a must
for DKG but I thought it is good to stay consistent. The identification is based
on DKG seed and message being signed. My first attempt was to use a session ID
but the session ID is computed separately for each attempt and we want to stop
the main loop, not just the given attempt.