Introduce broadcast channel retransmission strategies
So far, the broadcast channel used a simple retransmission strategy that
triggered a new retransmission on every new tick. Such an aggressive approach
turned to be not suitable for all use cases. Specifically, this approach tends
to cause network congestion if a big amount of messages is retransmitted for
a long time. To make the retransmission mechanism more flexible, here we
introduce the way to customize it through strategies. We are also add an
implementation of a backoff strategy that decreases its retransmission
frequency over time in an exponential way.