Integrate tBTCv2 rewards generation script
With this integration the way that 'bonus' and 'ongoing' rewards are
calculated is changing also.

Now, instead of calculate the rewards from the start of rewards period
(Jun 1st 2022), which requires heavy queries to subgraph and etherscan,
and long processing times, the rewards are calculated using the previous
distribution. Only the stake data of the new distribution period is
queried, and the rewards calculated using only this period are added to
the previous reward amounts.