Handling October 17th scenario for the uptime requirement
If a node is set before Oct17, we calculate the rewards if it was set on
Oct1. It still needs to meet a 96% uptime req though. E.g. if a node was
set on Oct15 it still has to meet 96% between Oct15-Oct31. If it meets
the reqs, then the rewards are distributed for the entire month. If not,
then no rewards.