Implement GetTransactionConfirmations
The function returns number of confirmations the transaction has.

It gets the transaction details and extracts the status data.

If the transaction is not found it retrurns an error.
If the transaction is not confirmed it returns `0`.

If the transaction is confirmed it compares the block number in which
transaction was included with the current block tip.

If the current block tip is the same number as transaction block number
then it retruns `1`.
If the current block tip is greater transaction block number it returns
the difference incremented by 1.

Examples:

latest block height = 2000
transaction block height = 2000
result = 1

latest block height = 2020
transaction block height = 2000
result = 21