Prepared signing protocol and member for parallel signing
We want to make the signing protocol more Bitcoin-friendly by allowing the
signing of multiple transaction inputs at the same time. Instead of triggering
multiple state machines, we will execute one progressing over signing phases
one by one, Each phase will be updating multiple TSS lib participants, one per
every message being signed.

In this commit, the signing member is updated to keep TSS party state separately,
for every message being signed. All protocol state execution functions received
a "composite" counterpart that is responsible for flattening composite messages
and calling internal processing functions, for every message being signed, in
the context of that message.

I have tried to abstract the flattening/composite code but that lead to quite
a complicated code with a lot of boilerplate. Landed with the current version
as the most readable one.

This commit is just another step on the path. The refactoring affects multiple
files so I am committing this set of changes as a logical unit of work but with
a failing build.